Revert "Another dependency the bbot missed!!!"
This reverts commit 161f7fd3bdd425167af9fe26fdc5373a2ff44c98.

Revert "Missed a file as part of checkin for r48186"

This reverts commit cff86beba5938209393a6c3bccced62a7f3ff36b.

Revert "Enable warning 4389 as an error on windows builds.  This will make"

This reverts commit c78936bcfc65b98edf288191d927a495b0364621.

TBR=mbelshe

Review URL: http://codereview.chromium.org/2253001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@48238 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/scoped_bstr_win_unittest.cc
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2010 The Chromium Authors. All rights reserved.
+// Copyright (c) 2006-2008 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -53,10 +53,10 @@ void BasicBstrTests() {
   EXPECT_TRUE(b2.ByteLength() == 100);
   EXPECT_TRUE(b2.Length() == 100 / sizeof(kTestString1[0]));
   lstrcpyW(static_cast<BSTR>(b2), kTestString1);
-  EXPECT_TRUE(lstrlen(b2) == static_cast<int>(test1_len));
+  EXPECT_TRUE(lstrlen(b2) == test1_len);
   EXPECT_TRUE(b2.Length() == 100 / sizeof(kTestString1[0]));
   b2.SetByteLen(lstrlen(b2) * sizeof(kTestString2[0]));
-  EXPECT_TRUE(lstrlen(b2) == static_cast<int>(b2.Length()));
+  EXPECT_TRUE(lstrlen(b2) == b2.Length());
 
   EXPECT_TRUE(b1.Allocate(kTestString2) != NULL);
   EXPECT_TRUE(b1.Length() == test2_len);
