+++ chrome_frame/chrome_launcher.cc
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2010 The Chromium Authors. All rights reserved.
+// Copyright (c) 2009 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -66,7 +66,7 @@ CommandLine* CreateLaunchCommandLine() {
 }
 
 void SanitizeCommandLine(const CommandLine& original, CommandLine* sanitized) {
-  size_t num_sanitized_switches = 0;
+  int num_sanitized_switches = 0;
   for (int i = 0; i < arraysize(kAllowedSwitches); ++i) {
     const char* current_switch = kAllowedSwitches[i];
     if (original.HasSwitch(current_switch)) {
